Refactor: ProfileImage 경로 수정
